Smart-System-for-Seaweed-Cultivation-and-Harvest-Management/requirements.txt
tensorflow>=2.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0
flask>=3.0.0
flask-cors>=4.0.0
opencv-python